+++ iterative_ilqr/control/ilqr_helper.py
@@ -160,68 +160,3 @@ def get_termianl_state_idex(ss_xcurv, Qfun, xt):
     norm = la.norm(diff, 1, axis=1)
     min_index = np.argmin(norm)
     return min_index
-
-
-def backward_pass(
-    xvar, uvar, x_terminal, dX, lamb, num_horizon, f_x, f_u, ilqr_param, obstacle, sys_param
-):
-    # cost derivation
-    l_u, l_uu, l_x, l_xx = get_cost_derivation(
-        uvar, dX, ilqr_param, num_horizon, xvar, obstacle, sys_param
-    )
-    # Value function at last timestep
-    matrix_Vx, matrix_Vxx = get_cost_final(
-        xvar, x_terminal, ilqr_param.matrix_Qlamb, obstacle, ilqr_param
-    )
-    # define control modification k and K
-    matrix_K = np.zeros((U_DIM, X_DIM, num_horizon))
-    matrix_k = np.zeros((U_DIM, num_horizon))
-    for idx_b in range(num_horizon - 1, -1, -1):
-        matrix_Qx = l_x[:, idx_b] + f_x[:, :, idx_b].T @ matrix_Vx
-        matrix_Qu = l_u[:, idx_b] + f_u[:, :, idx_b].T @ matrix_Vx
-        matrix_Qxx = l_xx[:, :, idx_b] + f_x[:, :, idx_b].T @ matrix_Vxx @ f_x[:, :, idx_b]
-        matrix_Quu = l_uu[:, :, idx_b] + f_u[:, :, idx_b].T @ matrix_Vxx @ f_u[:, :, idx_b]
-        matrix_Qux = f_u[:, :, idx_b].T @ matrix_Vxx @ f_x[:, :, idx_b]
-        # Improved Regularization
-        matrix_Quu_evals, matrix_Quu_evecs = np.linalg.eig(matrix_Quu)
-        matrix_Quu_evals[matrix_Quu_evals < 0] = 0.0
-        matrix_Quu_evals += lamb
-        matrix_Quu_inv = np.dot(
-            matrix_Quu_evecs, np.dot(np.diag(1.0 / matrix_Quu_evals), matrix_Quu_evecs.T)
-        )
-        # Calculate feedforward and feedback terms
-        matrix_k[:, idx_b] = -matrix_Quu_inv @ matrix_Qu
-        matrix_K[:, :, idx_b] = -matrix_Quu_inv @ matrix_Qux
-        # Update value function for next time step
-        matrix_Vx = matrix_Qx - matrix_K[:, :, idx_b].T @ matrix_Quu @ matrix_k[:, idx_b]
-        matrix_Vxx = matrix_Qxx - matrix_K[:, :, idx_b].T @ matrix_Quu @ matrix_K[:, :, idx_b]
-    return matrix_k, matrix_K
-
-
-def forward_pass(
-    xvar, uvar, x_terminal, ilqr_param, timestep, num_horizon, matrix_k, matrix_K, sys_param
-):
-    xvar_new = np.zeros((X_DIM, num_horizon + 1))
-    xvar_new[:, 0] = xvar[:, 0]
-    uvar_new = np.zeros((U_DIM, num_horizon))
-    cost_new = 0
-    for idx_f in range(num_horizon):
-        uvar_new[:, idx_f] = (
-            uvar[:, idx_f]
-            + matrix_k[:, idx_f]
-            + matrix_K[:, :, idx_f] @ (xvar_new[:, idx_f] - xvar[:, idx_f])
-        )
-        uvar_new[0, idx_f] = np.clip(uvar_new[0, idx_f], -sys_param.a_max, sys_param.a_max)
-        uvar_new[1, idx_f] = np.clip(uvar_new[1, idx_f], -sys_param.delta_max, sys_param.delta_max)
-        xvar_new[:, idx_f + 1] = kinetic_bicycle(xvar_new[:, idx_f], uvar_new[:, idx_f], timestep)
-        l_state_temp = (
-            (xvar_new[:, idx_f] - x_terminal).T
-            @ ilqr_param.matrix_Q
-            @ (xvar_new[:, idx_f] - x_terminal)
-        )
-        l_ctrl_temp = uvar_new[:, idx_f].T @ ilqr_param.matrix_R @ uvar_new[:, idx_f]
-        cost_new = cost_new + l_state_temp + l_ctrl_temp
-    dx_terminal_temp = xvar_new[:, -1] - x_terminal.T
-    dx_terminal_temp = dx_terminal_temp.reshape(X_DIM)
-    cost_new = cost_new + dx_terminal_temp.T @ ilqr_param.matrix_Qlamb @ dx_terminal_temp
-    return xvar_new, uvar_new, cost_new
